fixed mobile support and some scss
