+++ assets/_styles.scss
@@ -7,7 +7,7 @@
 }
 
 #p4{
-    background-color:$gold;
+    background-color: $gold;
 }
 
 #p5 {
