+++ assets/index.css
@@ -7,6 +7,12 @@ body {
   font-size: 3vw;
 }
 
+@media only screen and (max-width: 600px) {
+  .child {
+    font-size: 5vw;
+  }
+}
+
 * {
   padding: 0;
   margin: 0;
@@ -25,8 +31,8 @@ html {
 .parent {
   display: -ms-grid;
   display: grid;
-  -ms-grid-columns: (50%)[2];
-      grid-template-columns: repeat(2, 50%);
+  -ms-grid-columns: 100vw;
+      grid-template-columns: 100vw;
   -ms-grid-rows: (100vh)[13];
       grid-template-rows: repeat(13, 100vh);
   grid-column-gap: 0px;
@@ -37,104 +43,104 @@ html {
   -ms-grid-row: 1;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 1 / 1 / 2 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 1 / 1 / 2 / 2;
 }
 
 #p2 {
   -ms-grid-row: 2;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 2 / 1 / 3 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 2 / 1 / 3 / 2;
 }
 
 #p3 {
   -ms-grid-row: 3;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 3 / 1 / 4 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 3 / 1 / 4 / 2;
 }
 
 #p4 {
   -ms-grid-row: 4;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 4 / 1 / 5 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 4 / 1 / 5 / 2;
 }
 
 #p5 {
   -ms-grid-row: 5;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 5 / 1 / 6 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 5 / 1 / 6 / 2;
 }
 
 #p6 {
   -ms-grid-row: 6;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 6 / 1 / 7 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 6 / 1 / 7 / 2;
 }
 
 #p7 {
   -ms-grid-row: 7;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 7 / 1 / 8 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 7 / 1 / 8 / 2;
 }
 
 #p8 {
   -ms-grid-row: 8;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 8 / 1 / 9 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 8 / 1 / 9 / 2;
 }
 
 #p9 {
   -ms-grid-row: 9;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 9 / 1 / 10 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 9 / 1 / 10 / 2;
 }
 
 #p10 {
   -ms-grid-row: 10;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 10 / 1 / 11 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 10 / 1 / 11 / 2;
 }
 
 #p11 {
   -ms-grid-row: 11;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 11 / 1 / 12 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 11 / 1 / 12 / 2;
 }
 
 #p12 {
   -ms-grid-row: 12;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 12 / 1 / 13 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 12 / 1 / 13 / 2;
 }
 
 #p13 {
   -ms-grid-row: 13;
   -ms-grid-row-span: 1;
   -ms-grid-column: 1;
-  -ms-grid-column-span: 2;
-  grid-area: 13 / 1 / 14 / 3;
+  -ms-grid-column-span: 1;
+  grid-area: 13 / 1 / 14 / 2;
 }
 
 section {
@@ -240,6 +246,33 @@ section {
 .sand {
   height: 150px;
   width: 100vw;
+}
+
+@media only screen and (max-width: 600px) {
+  .tree {
+    height: 150px;
+    width: 150px;
+  }
+  .heart {
+    height: 100px;
+    width: 100px;
+  }
+  .cloud {
+    height: 50px;
+    width: 50px;
+  }
+  .sand {
+    height: 50px;
+  }
+  #p8-clouds > img, #p7-clouds > img {
+    margin-left: 1rem;
+  }
+  .flex-row {
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+        -ms-flex-direction: column;
+            flex-direction: column;
+  }
 }
 
 #p3, #p4, #p5, #p7, #p8, #p9, #p11 {
